プラクティスの削除機能実装

+++ app/src/main/java/jp/ceed/kart/settings/ui/practice/fragment/PracticeDetailFragment.kt
@@ -13,6 +13,7 @@ import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import jp.ceed.kart.settings.R
 import jp.ceed.kart.settings.databinding.FragmentPracticeDetailBinding
+import jp.ceed.kart.settings.ui.common.fragment.DeleteConfirmDialogFragment
 import jp.ceed.kart.settings.ui.practice.adapter.PracticeDetailAdapter
 import jp.ceed.kart.settings.ui.practice.viewModel.PracticeDetailFragmentViewModel
 import jp.ceed.kart.settings.ui.util.UiUtil
@@ -62,13 +63,13 @@ class PracticeDetailFragment: Fragment() {
     private fun onViewModelEvent(eventContent: PracticeDetailFragmentViewModel.EventContent){
         when(eventContent.eventType){
             PracticeDetailFragmentViewModel.EventType.CLICK_DELETE_DIALOG -> {
-                DeleteSessionDialogFragment { button ->
+                DeleteConfirmDialogFragment { _, button ->
                     when(button){
                         DialogInterface.BUTTON_POSITIVE -> {
                             viewModel.deleteSession(eventContent.value)
                         }else -> {}
                     }
-                }.show(childFragmentManager, DeleteSessionDialogFragment.TAG)
+                }.show(childFragmentManager, DeleteConfirmDialogFragment.TAG)
             }
             PracticeDetailFragmentViewModel.EventType.EDIT_MODE_COMPLETED -> {
                 UiUtil(requireContext()).hideKeyboard(binding.root)
